requirements.txt
ultralytics>=8.0.0
opencv-python
numpy
pytesseract
arabic-reshaper
python-bidi
